% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MicroarrayNormalize.R
\name{ma.normalize}
\alias{ma.normalize}
\title{Microarray normalization}
\usage{
ma.normalize(config)
}
\arguments{
\item{config}{Character string consisting of the path to the configuration
file generated using the \code{write.yaml.config} function or parsed
configuration list associated with a microarray experiment.}
}
\value{
Returns an expression set object that contains the microarray data at
the level of probes.
}
\description{
Background correct and normalize a microarray experiment.
}
\details{
Reads in all the samples associated with a microarray experiment according to
the options in the config and combines them using the Robust Multichip
Average (RMA) algorithm, which background corrects, quantile normalizes, and
summarizes the data using median polish at the level of probes in an
expression set object (eset). If the option \code{normalization: no} is set
in the config, an expression set is returned which has neither been
background corrected nor normalized, a feature which is useful for quality
assessment.
}
\examples{
if(require(madeData))
{
  config <- system.file("extdata", "config.yaml", package = "madeData")
  ma.normalize(config)
}

}
\references{
Irizarry, Rafael A., Bridget Hobbs, Francois Collin, Yasmin D.
Beazer-Barclay, Kristen J. Antonellis, Uwe Scherf, and Terence P. Speed.
"Exploration, normalization, and summaries of high density oligonucleotide
array probe level data." \emph{Biostatistics} 4, no. 2 (2003): 249-264.

Carvalho, Benilton S., and Rafael A. Irizarry. "A framework for
oligonucleotide microarray preprocessing." \emph{Bioinformatics} 26, no. 19
(2010): 2363-2367.
}
\seealso{
\code{\link{write.yaml.config}} to generate the configuration file,
and \code{\link{ma.summarize}} to convert an expression set object into a
list of genes and their associated log-fold changes and statistical values.
}

