% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MicroarrayPipeline.R
\name{ma.pipeline}
\alias{ma.pipeline}
\title{Microarray pipeline}
\usage{
ma.pipeline(config)
}
\arguments{
\item{config}{Character string consisting of the path to the configuration
file generated using the \code{write.yaml.config} function or parsed
configuration list associated with a microarray experiment.}
}
\description{
Run the full microarray pipeline.
}
\details{
Runs the entire microarray pipeline according to the options set in the
config. The pipeline consists of quality assessment, normalization,
summarization, and creation of the final report.
}
\seealso{
\code{\link{write.yaml.config}} to generate the configuration file,
\code{\link{ma.normalize}} to perform normalization,
\code{\link{ma.summarize}} to perform gene level summarization, and
\code{\link{write.report}} to generate the final report including the quality
assessment.
}

