% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGroupFile.R
\name{read.group.file}
\alias{read.group.file}
\title{Read group file}
\usage{
read.group.file(config)
}
\arguments{
\item{config}{Character string consisting of the path to the configuration
file generated using the \code{write.yaml.config} function or parsed
configuration list associated with a microarray experiment.}
}
\value{
Returns a modified configuration with group data.
}
\description{
Read and validate the group file associated with a microarray experiment.
}
\details{
\strong{Note}: This function is mostly for internal use. Users should use
\code{read.yaml.config} instead because it calls this function by default.

The group file describes which samples belong to which groups or conditions,
for example normal controls and experimental cases. Groups must be assigned
prior to launching the analysis. Groups are compared according to the option
\code{contrast_groups} in the configuration and can be defined at three
different levels:

\describe{
   \item{\code{"dirs" }}{Each directory containing samples can be assigned a
   group.}
   \item{\code{"files"}}{Each individual sample can be assigned a group.}
   \item{\code{"eset" }}{Each sample specified in an expression set can be
   assigned a group.}
 }
}
\examples{
if(require(madeData))
{
  config <- system.file("extdata", "config.yaml", package = "madeData")
  read.group.file(config)
}

}
\seealso{
\code{\link{write.yaml.config}} to generate the configuration file
and \code{\link{read.yaml.config}} to read and validate it.
}

