% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ReadYamlConfig.R
\name{read.yaml.config}
\alias{read.yaml.config}
\title{Read YAML config}
\usage{
read.yaml.config(config, checkGroups = TRUE, getGroups = TRUE)
}
\arguments{
\item{config}{Character string consisting of the path to the configuration
file generated using the \code{write.yaml.config} function or parsed
configuration list associated with a microarray experiment.}

\item{checkGroups}{logical indicating whether the group file must exist when
doing the validation.}

\item{getGroups}{logical indicating whether the data in the group file must
also be validated and returned in the configuration.}
}
\description{
Read and validate the configuration associated with a microarray experiment.
}
\details{
Each configuration option is checked for incongruities, including whether or
not it is missing, its type, and its value.
}
\seealso{
\code{\link{write.yaml.config}} to generate the configuration file
and \code{\link{read.group.file}} to read and validate the group file.
}

