% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WriteReport.R
\name{write.report}
\alias{write.report}
\title{Write microarray report}
\usage{
write.report(config, eset, tf)
}
\arguments{
\item{config}{Character string consisting of the path to the configuration
file generated using the \code{write.yaml.config} function or parsed
configuration list associated with a microarray experiment.}

\item{eset}{Expression set object describing microarray experiment at the
level of probes.}

\item{tf}{A list of top tables for each comparison of interest and their
genes, associated log-fold changes, and statistical values.}
}
\description{
Generate the final microarray report.
}
\details{
Generates the microarray report summarizing the information in the
experiment. The information in the report depends on the config and the
packages available.
}

