% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WriteYamlConfig.R
\name{write.yaml.config}
\alias{write.yaml.config}
\title{Write YAML configuration}
\usage{
write.yaml.config(analysisDir, groupBy = NULL, ...)
}
\arguments{
\item{analysisDir}{Character string consisting of the path to the analysis
directory.}

\item{groupBy}{Must be one of \code{"dirs"}, \code{"files"}, or
\code{"eset"}.}

\item{...}{Additional options to pass to the function.}
}
\description{
Generate the configuration file describing a microarray experiment.
}
\details{
Generates the configuration and group file which describe a reproducible
microarray experiment. The files are created in the analysis directory using
the \code{analysisDir} option. The group file describes which samples belong
to which groups or conditions, for example normal controls and experimental
cases. The user must specify at which level they are grouping their samples
using the \code{groupBy} option:

\describe{
   \item{\code{"dirs" }}{Each directory containing samples can be assigned a
   group. This option is useful if there are many samples in the experiment.
   If this option is specified, before running this function the analysis
   directory must already contain subdirectories of samples.}
   \item{\code{"files"}}{Each individual sample can be assigned a group. If
   this option is specified, before running this function the analysis
   directory must already contain samples.}
   \item{\code{"eset" }}{Each sample specified in an expression set can be
   assigned a group. If this option is specified, the expression set must
   also be passed as an additional argument to the function.}
 }

Other configuration options can also be specified as additional arguments,
otherwise defaults will be used:

\describe{
   \item{\code{qvalue}}{A number between 0 and 1 which corrects the pvalue
   for multiple comparisons and describes the minimum false discovery rate
   (FDR) at which a test may be called significant: the default is
   \code{0.05}.}
   \item{\code{save.intermediates}}{A logical indicating whether or not
   intermediate outputs should be saved: the default is \code{FALSE}.}
   \item{\code{quality.assessment}}{A logical indicating whether or not
   quality assessment should be performed: the default is \code{TRUE}.}
   \item{\code{normalization}}{Either \code{"rma"}, \code{TRUE}, or
   \code{FALSE}, indicating what sort of background correction and
   normalization should be performed, if any. Setting this value to
   \code{FALSE} is useful for quality assessment, since the expression set
   will not be background corrected or normalized: the default is
   \code{"rma"}.}
   \item{\code{summarization}}{A logical indicating whether or not gene level
   summarization should be performed: the default is \code{TRUE}.}
   \item{\code{write.report}}{A logical indicating whether or not a report of
   the experiment should be created: the default is \code{TRUE}.}
   \item{\code{group.file}}{A character string describing the name to give
   the group file: the default is to name the file \code{"groups.txt"}.}
   \item{\code{contrast.groups}}{A character string of one or more formulas
   separated by a comma describing how the different groups should be
   compared: by default groups \code{"A"}, \code{"B"}, and \code{"C"} are
   created and are each compared to each other.}
   \item{\code{eset}}{An expression set object describing the microarray
   experiment at the level of probes. This option should only be used if the
   \code{groupBy} option is set to \code{"eset"}, and thus has no default.}
   \item{\code{groups.df}}{A \code{data.frame} indicating which samples
   belong to which groups.

This option has no default: it can be used to programmatically set the
   groups for each sample ahead of time, otherwise the user will need to
   modify the group file after the configuration has been created.}
 }
}
\seealso{
\code{\link{read.yaml.config}} to read and validate the
configuration.
}

