% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteReport.R
\name{write.report}
\alias{write.report}
\title{Write microarray report}
\usage{
write.report(config, eset, results)
}
\arguments{
\item{config}{Character string consisting of the path to the configuration
file generated using the \code{write.yaml.config} function or parsed
configuration list associated with a microarray experiment.}

\item{eset}{Expression set object describing microarray experiment at the
level of probes.}

\item{results}{A summary of all the data in the expression set for each
group comparison generated using the \code{ma.summarize} function.}
}
\value{
A character string consisting of the path to the microarray report.
}
\description{
Generate the final microarray report.
}
\details{
Generates the microarray report summarizing the information in the
experiment. The information in the report depends on the config and the
packages available.
}
\examples{
write.report(config, eset, results)

}

